+++ src/lib/djerba/plugins/sample/plugin.py
@@ -15,20 +15,24 @@ import djerba.core.constants as core_constants
 from djerba.util.subprocess_runner import subprocess_runner
 from djerba.util.render_mako import mako_renderer
 from djerba.util.logger import logger
+from djerba.util.validator import path_validator
 
 
 class main(plugin_base):
 
     PLUGIN_VERSION = '1.0.0'
-    QCETL_CACHE = "/scratch2/groups/gsi/production/qcetl_v1"
+    QCETL_CACHE_DEFAULT = "/scratch2/groups/gsi/production/qcetl_v1"
+    QCETL_CACHE_KEY = 'qcetl_cache'
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         if find_spec('gsiqcetl')==None:
-            msg = "GSI-QC-ETL not found: Coverage and callability must "+\
-                "be specified manually. GSI-QC_ETL is developed and run "+\
-                "internally at OICR; it is not available externally."
-            self.logger.warn(msg)
+            # gsiqcetl Python package is not on the PYTHONPATH
+            warning = "GSI-QC-ETL API not found: Coverage and callability must "+\
+                "be specified manually in the INI file. (The GSI-QC-ETL cache and "+\
+                "associated Python package are "+\
+                "internal to OICR and not available externally.)"
+            self.logger.warn(warning)
             self.gsiqcetl_OK = False
         else:
             try:
@@ -38,7 +42,6 @@ class main(plugin_base):
             except ImportError as err:
                 msg = 'QC-ETL import failure! Try checking python versions'
                 raise RuntimeError(msg) from err
-
 
     def specify_params(self):
         discovered = [
@@ -54,11 +57,12 @@ class main(plugin_base):
         for key in discovered:
             self.add_ini_discovered(key)
         
-        # Default parameters for priorities
+        # Default parameters
         self.set_ini_default('configure_priority', 100)
         self.set_ini_default('extract_priority', 500)
         self.set_ini_default('render_priority', 500)
         self.set_ini_default(constants.CALLABILITY_WARNING, False)
+        self.set_ini_default(self.QCETL_CACHE_KEY, self.QCETL_CACHE_DEFAULT)
 
     def configure(self, config):
         config = self.apply_defaults(config)
@@ -80,7 +84,6 @@ class main(plugin_base):
         wrapper = self.fill_param_if_null(wrapper, constants.PURITY, "purity_ploidy.json")
         wrapper = self.fill_param_if_null(wrapper, constants.PLOIDY, "purity_ploidy.json")
 
-
         # Get tumour_id and donor
         for key in [core_constants.TUMOUR_ID, constants.DONOR]:
             if wrapper.my_param_is_null(key):
@@ -92,22 +95,26 @@ class main(plugin_base):
                     self.logger.error(msg)
                     raise RuntimeError(msg)
 
-        # Fetch tumour_id and donor
-        donor = config[self.identifier][constants.DONOR]
-        tumour_id = config[self.identifier][core_constants.TUMOUR_ID]
+        # Store config parameters as variables for convenience
+        donor = wrapper.get_my_string(constants.DONOR)
+        tumour_id = wrapper.get_my_string(core_constants.TUMOUR_ID)
         ignore_warning = wrapper.get_my_boolean(constants.CALLABILITY_WARNING)
+        cache_path = wrapper.get_my_string(self.QCETL_CACHE_KEY)
 
         # SECOND PASS: Get files based on input parameters
         if self.gsiqcetl_OK:
+            etl_cache = self.get_qcetl_cache(cache_path)
             if wrapper.my_param_is_null(constants.CALLABILITY):
-                self.logger.debug("Fetching callability from GSI-QC_ETL")
-                wrapper.set_my_param(constants.CALLABILITY, self.fetch_callability_etl_data(donor, tumour_id, ignore_warning))
+                self.logger.debug("Fetching callability from GSI-QC-ETL")
+                callability = self.fetch_callability_etl_data(etl_cache, donor, tumour_id, ignore_warning)
+                wrapper.set_my_param(constants.CALLABILITY, callability)
             if wrapper.my_param_is_null(constants.COVERAGE):
-                self.logger.debug("Fetching coverage from GSI-QC_ETL")
-                wrapper.set_my_param(constants.COVERAGE, self.fetch_coverage_etl_data(donor, tumour_id))
+                self.logger.debug("Fetching coverage from GSI-QC-ETL")
+                coverage = self.fetch_coverage_etl_data(etl_cache, donor, tumour_id)
+                wrapper.set_my_param(constants.COVERAGE, coverage)
         else:
             msg = "GSI-QC-ETL not available, omitting coverage/callability fetch"
-            self.logger.debug(msg)
+            self.logger.info(msg)
         return wrapper.get_config()
 
     def extract(self, config):
@@ -139,8 +146,13 @@ class main(plugin_base):
         renderer = mako_renderer(self.get_module_dir())
         return renderer.render_name('sample_template.html', data)
 
-    def fetch_callability_etl_data(self, donor, tumour_id, ignore_warning):
-        etl_cache = QCETLCache(self.QCETL_CACHE)
+    def get_qcetl_cache(self, cache_path):
+        val = path_validator(self.log_level, self.log_path)
+        val.validate_input_dir(cache_path)
+        etl_cache = QCETLCache(cache_path)
+        return etl_cache
+
+    def fetch_callability_etl_data(self, etl_cache, donor, tumour_id, ignore_warning):
         cached_callabilities = etl_cache.mutectcallability.mutectcallability
         columns_of_interest = gsiqcetl.column.MutetctCallabilityColumn
         # Note: donor and tumour ID are both not unique, but together are unique. Filter on both.
@@ -168,8 +180,7 @@ class main(plugin_base):
             self.logger.error(msg)
             raise MissingQCETLError(msg)
         
-    def fetch_coverage_etl_data(self, donor, tumour_id):
-        etl_cache = QCETLCache(self.QCETL_CACHE)
+    def fetch_coverage_etl_data(self, etl_cache, donor, tumour_id):
         cached_coverages = etl_cache.bamqc4merged.bamqc4merged
         columns_of_interest = gsiqcetl.column.BamQc4MergedColumn
         # Note: donor and tumour ID are both not unique, but together are unique. Filter on both.